Move save dialog to component
